+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added username param to !coords command
 - Added a rod timer sound that gets more high-pitched the lower the timer is
 - Added Bobbin' Time Display feature to track nearby bobbers and show the calculated buff percentage
+- Added an alert when attempting to fish with a rod missing a Hook, Line, or Sinker
 
 ### Changes
 - Made the bestiary display show in alphabetical order